Split config into several submodules
